+++ data-filter/lib/filter/filters/options.filter.ts
@@ -8,36 +8,41 @@ import { FilterOperatorTypes } from "../operators";
 import { FilterType } from "../type";
 import { BaseFilterDefinition, Filter, FilterCondition, FilterConditionRule } from "./filter";
 
-export interface RadioFilterOption {
+export interface OptionsFilterOption {
     key: string;
     value?: unknown;
     operator?: FilterOperatorTypes;
     condition?: FilterCondition;
 }
 
-export interface RadioFilterDefinition {
-    options: RadioFilterOption[];
+export type OptionsFilterSelectionMode = "radio" | "select";
+
+export interface OptionsFilterDefinition {
+    selectionMode: OptionsFilterSelectionMode;
+    options: OptionsFilterOption[];
 }
 
-export interface RadioFilterOptionConfiguration {
+export interface OptionsFilterOptionConfiguration {
     key: string;
     name?: string;
 }
 
-export interface RadioFilterConfigurationModel extends FilterBaseConfigurationModel {
-    options: RadioFilterOptionConfiguration[];
+export interface OptionsFilterConfigurationModel extends FilterBaseConfigurationModel {
+    selectionMode: OptionsFilterSelectionMode;
+    options: OptionsFilterOptionConfiguration[];
 }
 
-export class RadioFilter extends Filter implements RadioFilterDefinition {
-    public type = FilterType.Radio;
+export class OptionsFilter extends Filter implements OptionsFilterDefinition {
+    public type = FilterType.Options;
     public operators = [FilterOperatorTypes.Equal];
-    public options: RadioFilterOption[];
+    public selectionMode!: OptionsFilterSelectionMode;
+    public options!: OptionsFilterOption[];
 
-    constructor(definition: BaseFilterDefinition & RadioFilterDefinition) {
+    constructor(definition: BaseFilterDefinition & OptionsFilterDefinition) {
         super(definition);
     }
 
-    public async getConfig<Request>(key: string, request: Request, user?: DataFilterUserModel): Promise<RadioFilterConfigurationModel> {
+    public async getConfig<Request>(key: string, request: Request, user?: DataFilterUserModel): Promise<OptionsFilterConfigurationModel | null> {
         const config = await super.getConfig(key, request, user);
 
         if (!config) {
@@ -46,10 +51,11 @@ export class RadioFilter extends Filter implements RadioFilterDefinition {
 
         return {
             ...config,
+            selectionMode: this.selectionMode,
             options: await Promise.all(this.options.map(async x => ({
                 key: x.key,
                 name: await this._translateService.getTranslation(
-                    user?.language,
+                    user?.language ?? "",
                     FilterUtils.getRadioOptionTranslationKey(key, x.key)
                 )
             })))
@@ -66,7 +72,7 @@ export class RadioFilter extends Filter implements RadioFilterDefinition {
         return this.getRulePaths(condition.rules, data);
     }
 
-    public async getWhereOptions(rule: QueryRuleModel): Promise<WhereOptions> {
+    public async getWhereOptions(rule: QueryRuleModel): Promise<WhereOptions | undefined> {
         const option = this.options.find(x => x.key === rule.value);
         if (!option) {
             return super.getWhereOptions(rule);
@@ -80,7 +86,7 @@ export class RadioFilter extends Filter implements RadioFilterDefinition {
             });
         }
 
-        const conditions = [];
+        const conditions: any[] = [];
         const where = option.condition.condition === "and" ? { [Op.and]: conditions } : { [Op.or]: conditions };
         this.generateRuleWhereOptions(option.condition, conditions, rule);
         conditions.push(
@@ -93,7 +99,7 @@ export class RadioFilter extends Filter implements RadioFilterDefinition {
         return where;
     }
 
-    public async getHavingOptions(rule: QueryRuleModel): Promise<WhereOptions> {
+    public async getHavingOptions(rule: QueryRuleModel): Promise<WhereOptions | undefined> {
         const option = this.options.find(x => x.key === rule.value);
         if (!option) {
             return super.getHavingOptions(rule);
@@ -114,14 +120,33 @@ export class RadioFilter extends Filter implements RadioFilterDefinition {
                 paths.push(...this.getRulePaths((rule as FilterCondition).rules, data));
             } else if ((rule as FilterConditionRule).path) {
                 if (paths.every((path) => path.path !== (rule as FilterConditionRule).path)) {
+                    const r = rule as FilterConditionRule;
+                    if (!r.path) {
+                        continue;
+                    }
+
                     paths.push({
-                        path: (rule as FilterConditionRule).path,
-                        where: FilterUtils.generateWhereConditions((rule as FilterConditionRule).where, data)
+                        path: r.path,
+                        where: FilterUtils.generateWhereConditions(r.where, data)
                     });
                 }
             }
         }
 
         return paths;
+    }
+}
+
+/**
+ * @deprecated: Use OptionsFilter instead
+ */
+export class RadioFilter extends OptionsFilter {
+    public type = FilterType.Radio;
+
+    constructor(definition: BaseFilterDefinition & Omit<OptionsFilterDefinition, "selectionMode">) {
+        super({
+            ...definition,
+            selectionMode: "radio"
+        });
     }
 }