+++ data-filter/lib/models/custom-attributes.model.ts
@@ -1,5 +1,4 @@
 import { ProjectionAlias } from "sequelize";
-import { GroupOption } from "sequelize/types/model";
 import { IncludeWhereModel } from "./include.model";
 import { PathModel } from "./path.model";
 
@@ -14,7 +13,7 @@ export interface CustomAttributesConfig<T extends CustomAttributesOptionConfig =
     type: string;
     config?: T;
 
-    transform(options?: object, path?: string): string | ProjectionAlias;
+    transform(options?: object, path?: string): string | ProjectionAlias | null;
     shouldGroupBy(): boolean;
 }