+++ data-filter/lib/scanners/sequelize-model.scanner.ts
@@ -13,7 +13,7 @@ export class SequelizeModelScanner {
         model: typeof Model,
         path: PathModel,
         additionalIncludes: IncludeModel[],
-        attributes?: FindAttributeOptions,
+        attributes?: FindAttributeOptions | null,
         ignoreAttributes = false
     ): IncludeOptions[] {
         const objects = path?.path?.split(".");
@@ -86,7 +86,7 @@ export class SequelizeModelScanner {
             );
         }
 
-        includes[0].include.push(...SequelizeUtils.reduceIncludes(addIncludes, ignoreAttributes));
+        includes[0].include?.push(...SequelizeUtils.reduceIncludes(addIncludes, ignoreAttributes));
 
         return result;
     }
@@ -144,13 +144,13 @@ export class SequelizeModelScanner {
         return result;
     }
 
-    public getOrder(model: typeof Model, orderObj: OrderModel): Order {
+    public getOrder(model: typeof Model, orderObj: OrderModel): Order | undefined {
         if (!orderObj?.column || !orderObj.direction) {
             return;
         }
 
         const values = orderObj.column.split(".");
-        const column = values.pop();
+        const column = values.pop() as string;
         for (const value of values) {
             const association = this.findAssociation(model, value);
             model = association.getAssociatedClass() as unknown as typeof Model;
@@ -173,12 +173,12 @@ export class SequelizeModelScanner {
 
     public getGroup(model: typeof Model, orderObj: OrderModel): string[] {
         if (!orderObj || !orderObj.column || orderObj.direction === "") {
-            return;
+            return [];
         }
 
         const group: string[] = [];
         const values = orderObj.column.split(".");
-        const column = values.pop();
+        const column = values.pop() as string;
         for (const value of values) {
             const association = this.findAssociation(model, value);
             model = association.getAssociatedClass() as unknown as typeof Model;
@@ -200,7 +200,7 @@ export class SequelizeModelScanner {
             return [];
         }
 
-        const column = values.pop();
+        const column = values.pop() as string;
         const result: IncludeOptions[] = [];
         let includes: IncludeOptions[] = result;
         let i = 0;
@@ -231,7 +231,7 @@ export class SequelizeModelScanner {
     }
 
     private findAssociation(model: typeof Model, obj: string): BaseAssociation<any, any> {
-        const associations = getAssociations(model.prototype);
+        const associations = getAssociations(model.prototype) ?? [];
 
         const association = associations.find((a) => a.getAs() === obj);
         if (!association) {